feat: add ErrorHandler
ErrorHandler is a class that will either throw the error OR log it to the console. The flag
indicating ErrorHandler should throw is provided via withGlobalSettings and it can only be set once
in a project. There is no "reset" unless you call withGlobalSettings again.

 The problem that
this feature attempts to solve it to allow consumers to NOT have exceptions thrown in development
mode which can be frustrating. In turns, this also attempts to solve a problem with Hot Module
Reload where sometimes Profiles isn't clear on "hot reload" leading to Duplicating Exception being
thrown.

 Default is ErrorHandler will throw exceptions, this is to keep the rest of automapper
default behavior in tact.

BREAKING CHANGE: ErrorHandler with shouldThrow set to false (using withGlobalSettings) calls
"console.warn" which isn't a terminating statement. Although, there is no "true" breaking changes
but if you run into any, please open an issue.